+++ app/system/modules/user/app/lib/permissions.js
@@ -6,10 +6,12 @@ module.exports = {
 
     created: function () {
 
+        this.Roles = this.$resource('api/user/role/:id');
+
         this.debounced = [];
 
-        this.saveCb = Vue.util.debounce(function(role) {
-            this.$resource('api/user/role/:id').save({ id: 'bulk' }, { roles: this.debounced }).then(function () {
+        this.saveCb = Vue.util.debounce(function() {
+            this.Roles.save({ id: 'bulk' }, { roles: this.debounced }).then(function () {
                 this.$notify('Permissions saved');
             });